BI-1524: Build OS X ODBC Plugins

+++ bld/bin/build.sh
@@ -4,7 +4,7 @@
 # shellcheck source=prepare-shell.sh
 . "$(dirname "$0")/prepare-shell.sh"
 
-# if boost isn't already present, download it
+# if boost isn't already present, download it.
 if [ ! -d "$BOOST_DIR" ]; then
     mkdir -p "$ARTIFACTS_DIR"
     rm -f "$BOOST_ARCHIVE"
@@ -19,42 +19,60 @@ if [ ! -d "$BOOST_DIR" ]; then
     rm "$BOOST_ARCHIVE"
 fi
 
-# clear the MYSQL_HOME_DIR
-rm -rf "$MYSQL_HOME_DIR"
+if [ -e "$MYSQL_HOME_DIR" ]; then
+    echo "cleaning $MYSQL_HOME_DIR"
+    rm -rf "$MYSQL_HOME_DIR"
+fi
+echo "creating $MYSQL_HOME_DIR tree"
 mkdir -p "$MYSQL_HOME_DIR/lib"
 mkdir -p "$MYSQL_HOME_DIR/include"
 mkdir -p "$MYSQL_HOME_DIR/bin"
 
-# clear BUILD_SRC_DIR
-rm -rf "$BUILD_SRC_DIR"
+if [ -e "$BUILD_SRC_DIR" ]; then
+    echo "cleaning $BUILD_SRC_DIR"
+    rm -rf "$BUILD_SRC_DIR" || true
+fi
+echo "creating $BUILD_SRC_DIR tree"
 mkdir -p "$BUILD_SRC_DIR"
 
-# copy mysql source into BUILD_SRC_DIR
-cp -r "$PROJECT_ROOT"/* "$BUILD_SRC_DIR"/ || true
-rm -rf "$BUILD_SRC_DIR"/bld
+# copy mysql source into BUILD_SRC_DIR.
+echo "copying $PROJECT_ROOT into $BUILD_SRC_DIR"
+for x in "$PROJECT_ROOT"/*; do
+    if [ ! "$(basename "$x")" = bld ]; then
+        echo "...copying $x to $BUILD_SRC_DIR"
+        cp -r "$x" "$BUILD_SRC_DIR"/
+    fi
+done
 
-# copy mongosql-auth source into BUILD_SRC_DIR
+# copy mongosql-auth source into BUILD_SRC_DIR.
 cp -r "$MONGOSQL_AUTH_ROOT"/src/mongosql-auth "$BUILD_SRC_DIR"/plugin/auth/
 cat "$MONGOSQL_AUTH_ROOT"/src/CMakeLists.txt >> "$BUILD_SRC_DIR"/plugin/auth/CMakeLists.txt
 cp "$MONGOSQL_AUTH_ROOT"/cmake/*.cmake "$BUILD_SRC_DIR"/cmake
 
-# clear the BUILD_DIR
+# clear the BUILD_DIR.
 rm -rf "$BUILD_DIR"
 mkdir -p "$BUILD_DIR"
 
-# run CMake in the BUILD_DIR
+# run CMake in the BUILD_DIR.
 cd "$BUILD_DIR"
 if [ -n "$CMAKE_GENERATOR" ]; then
+    # shellcheck disable=SC2086
     cmake "$BUILD_SRC_DIR" -G "$CMAKE_GENERATOR" $CMAKE_ARGS
 else
+    # shellcheck disable=SC2086
     cmake "$BUILD_SRC_DIR" $CMAKE_ARGS
 fi
 
-# build mysqlclient and the unit test binary
+# build mysqlclient and the unit test binary.  the following eval statements
+# need to be unquoted because we want the shell to split on white space.
+# Shellcheck wants us to qute references, so we need to supress the warnings
+# for SC2086.
+# shellcheck disable=SC2086
 eval $BUILD
+# shellcheck disable=SC2086
 eval $BUILD_UNIT_TESTS
 
-# copy artifacts needed to build ODBC driver into MYSQL_HOME_DIR
+# copy artifacts needed to build ODBC driver into MYSQL_HOME_DIR.
 mysqlclient='libmysqlclient.a'
 if [ "$OS" = 'Windows_NT' ]; then
     mysqlclient='Release/mysqlclient.lib'
